small refactoring, addressing comments
